Release 2.0 Исправлена ошибка с нехваткой памяти

+++ app/src/main/java/com/mcal/disassembler/nativeapi/DisassemblerDumper.java
@@ -5,10 +5,10 @@ public class DisassemblerDumper
 	public static native void load(String path);
 	public static native boolean hasFile(String path);
 	public static native String getNameAt(long pos);
-	public static native String getDemangledNameAt(int pos);
-	public static native short getTypeAt(int pos);
-	public static native short getBindAt(int pos);
-	public static native int getSize();
+	public static native String getDemangledNameAt(long pos);
+	public static native int getTypeAt(long pos);
+	public static native int getBindAt(long pos);
+	public static native long getSize();
 	public static native String demangle(String name);
 	public static native String demangleOnly(String name);
 }
